Make NotificationMessage an Interface (#63)
* Extract INotification from NotificationMessage and rename fields.

* Use interface more broadly, instead of the concrete type.

* INotification -> NotificationMessage

* mCustomData -> mData
