+++ notification-hubs-test-app-refresh/src/main/java/com/example/notification_hubs_test_app_refresh/ui/main/NotificationDisplayAdapter.java
@@ -91,8 +91,8 @@ public class NotificationDisplayAdapter extends RecyclerView.Adapter<Notificatio
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
         NotificationMessage entry = mNotifications.get(position);
         holder.mTitle.setText(entry.getTitle());
-        holder.mBody.setText(entry.getMessage());
-        holder.mDataCardinality.setText(String.valueOf(entry.getCustomData().size()));
+        holder.mBody.setText(entry.getBody());
+        holder.mDataCardinality.setText(String.valueOf(entry.getData().size()));
     }
 
     /**
