+++ notification-hubs-test-app-refresh/src/main/java/com/example/notification_hubs_test_app_refresh/ui/main/NotificationListFragment.java
@@ -2,7 +2,6 @@ package com.example.notification_hubs_test_app_refresh.ui.main;
 
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
-import androidx.lifecycle.ViewModelProviders;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -13,7 +12,6 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -22,6 +20,7 @@ import android.widget.Toast;
 import com.example.notification_hubs_test_app_refresh.NotificationDetailActivity;
 import com.example.notification_hubs_test_app_refresh.R;
 import com.microsoft.windowsazure.messaging.notificationhubs.NotificationMessage;
+import com.microsoft.windowsazure.messaging.notificationhubs.BasicNotificationMessage;
 
 import java.util.HashMap;
 import java.util.List;
@@ -40,12 +39,12 @@ public class NotificationListFragment extends Fragment {
         super.onCreate(savedInstanceState);
         mViewModel = new ViewModelProvider(this).get(NotificationListViewModel.class);
 
-        mViewModel.addNotification(new NotificationMessage("Puppies for Adoption!", "6 puppies born Sunday, each could be yours for just $100,000.00.", new HashMap<String,String>()));
+        mViewModel.addNotification(new BasicNotificationMessage("Puppies for Adoption!", "6 puppies born Sunday, each could be yours for just $100,000.00.", new HashMap<String,String>()));
 
         Map<String, String> catData = new HashMap<>();
         catData.put("feline", "yes");
         catData.put("fun", "no");
-        mViewModel.addNotification(new NotificationMessage("Kittens for pickup!", "Please take them, we can't handle it!", catData));
+        mViewModel.addNotification(new BasicNotificationMessage("Kittens for pickup!", "Please take them, we can't handle it!", catData));
     }
 
     @Override
@@ -66,8 +65,8 @@ public class NotificationListFragment extends Fragment {
         notificationDisplayAdapter.setClickListener(message -> {
             Intent i  = new Intent(this.getActivity(), NotificationDetailActivity.class);
             i.putExtra(NotificationDetailActivity.INTENT_TITLE_KEY, message.getTitle());
-            i.putExtra(NotificationDetailActivity.INTENT_BODY_KEY, message.getMessage());
-            for (Map.Entry<String, String> row : message.getCustomData().entrySet()) {
+            i.putExtra(NotificationDetailActivity.INTENT_BODY_KEY, message.getBody());
+            for (Map.Entry<String, String> row : message.getData().entrySet()) {
                 i.putExtra(row.getKey(), row.getValue());
             }
             startActivity(i);
