+++ notification-hubs-test-app-refresh/src/main/java/com/example/notification_hubs_test_app_refresh/ui/main/NotificationListViewModel.java
@@ -1,22 +1,15 @@
 package com.example.notification_hubs_test_app_refresh.ui.main;
 
-import android.content.Context;
-import android.widget.Toast;
-
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-import com.example.notification_hubs_test_app_refresh.R;
+import com.microsoft.windowsazure.messaging.notificationhubs.NotificationMessage;
 import com.microsoft.windowsazure.messaging.notificationhubs.NotificationHub;
 import com.microsoft.windowsazure.messaging.notificationhubs.NotificationListener;
-import com.microsoft.windowsazure.messaging.notificationhubs.NotificationMessage;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
 
 public class NotificationListViewModel extends ViewModel {
     private final MutableLiveData<List<NotificationMessage>> mNotifications;
